Restyle markdown help formatting to not use background colors
While there, colorize the "about" section of `--help`.

Fixes #450; superseedes #452.
