+++ edgedb-cli-md/.gitignore
@@ -1,0 +1,3 @@
+/Cargo.lock
+/target
+
